split components / helpers into files
