+++ src/lib/openai.ts
@@ -1,0 +1,11 @@
+type ChatRole = 'user' | 'assistant' | 'system' | 'tool'
+
+interface ChatMessage {
+    role: ChatRole
+    content: string
+}
+
+export type {
+    ChatRole,
+    ChatMessage
+}
